% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_hrv_frequence.R
\name{compute_hrv_frecuence}
\alias{compute_hrv_frecuence}
\title{Función para el análisis espectral de la variabilidad de la frecuencia cardíaca en el dominio de la frecuencia}
\usage{
compute_hrv_frecuence(t1, b1, resample_index = 125, fs = 1000)
}
\arguments{
\item{t1}{Vector numérico. Tiempos acumulados (en milisegundos) de los latidos válidos.}

\item{b1}{Vector numérico. Intervalos RR (en milisegundos) correspondientes a los tiempos \code{t1}.}

\item{resample_index}{resample_index Intervalo de remuestreo deseado (en milisegundos). Por defecto, 125 ms (≈ 8 Hz).}

\item{fs}{Frecuencia de muestreo original del ECG (en Hz). Por defecto son 1000 Hz.}
}
\value{
Lista con:
\describe{
\item{P_ULF}{Potencia en la banda ultra baja frecuencia (<0.003 Hz).}
\item{P_VLF}{Potencia en la banda muy baja frecuencia (0.003–0.04 Hz).}
\item{P_LF}{Potencia en la banda baja frecuencia (0.04–0.15 Hz).}
\item{P_HF}{Potencia en la banda alta frecuencia (0.15–0.4 Hz).}
\item{P_TP}{Potencia total (0–0.4 Hz).}
\item{P_TPP}{Potencia total de la señal autocorrelada.}
\item{LF_HF_ratio}{Cociente entre bandas LF y HF.}
\item{fdatx}{Magnitudes del espectro de frecuencia (FFT).}
\item{freq}{Vector de frecuencias en Hz correspondiente a \code{fdatx}.}
}
}
\description{
La función \code{compute_hrv_frecuence()} realiza un análisis espectral de la variabilidad de la frecuencia cardíaca (HRV)
usando la transformada rápida de Fourier (FFT) para calcular la potencia en diferentes bandas de frecuencia estandarizadas
asociadas a componentes del sistema nervioso autónomo.
}
\section{Objetivo principal}{

\itemize{
\item Obtener potencias espectrales en bandas ULF, VLF, LF y HF.
\item Calcular el cociente LF/HF como estimador del balance simpático-parasimpático.
}
}

\section{Detalles}{

\enumerate{
\item Interpolación de la señal RR sobre una malla equiespaciada mediante \code{interp1()} tipo PCHIP.
\item Eliminación de la media para centrar la señal (variabilidad pura).
\item Cálculo de autocorrelación para suavizar y estabilizar el espectro.
\item Aplicación de ventana de Hann para mitigar efectos de borde.
\item Cálculo de la FFT y conversión a espectro de potencia.
\item Integración espectral por bandas estándar: ULF, VLF, LF, HF.
}
}

\section{Requisitos}{

Esta función requiere el paquete \pkg{signal} para las funciones \code{interp1()} y \code{conv()}.
}

\examples{
\dontrun{
# Simulación de señal RR:
t <- seq(0, 60000, by = 1000)  # Tiempos en ms (1 latido por segundo)
rr <- rep(1000, length(t))     # Intervalos RR constantes de 1000 ms

result <- HRVf2_R(t, rr)
plot(result$freq, result$fdatx, type = "l", xlab = "Frecuencia (Hz)", ylab = "Amplitud")
}

}
\seealso{
\code{\link[signal]{interp1}} para interpolación cúbica, y \code{\link{compute_hrv_time}} para análisis HRV en el dominio del tiempo.
}
